refactor(fetch): move common fetch logic to abstract FetchBase class
