+++ finalynx/fetch/__init__.py
@@ -15,8 +15,8 @@ import sys
 
 import finary_uapi.constants  # type: ignore # Set the credentials and cookies constants to a full path
 
-finary_uapi.constants.CREDENTIAL_FILE = os.path.join(os.path.dirname(__file__), "credentials.json")
-finary_uapi.constants.COOKIE_FILENAME = os.path.join(os.path.dirname(__file__), "localCookiesMozilla.txt")
+finary_uapi.constants.CREDENTIAL_FILE = os.path.join(os.path.dirname(__file__), "finary_credentials.json")
+finary_uapi.constants.COOKIE_FILENAME = os.path.join(os.path.dirname(__file__), "finary_cookies.txt")
 
 from .fetch_finary import FetchFinary
 from ..console import console
